EXTI GPIO Driver
Display Message on LCD when Rising edge of GPIOB Pin 9 happened

+++ U7_MCU_Peripherals/L4_EXTI_GPIO_Driver/STM32_F103C6_Drivers/inc/STM32F103x6_GPIO_Driver.h
@@ -117,7 +117,7 @@ uint16_t MCAL_GPIO_ReadPort (GPIO_TypeDef* GPIOx );
 
 //Write APIs
 void MCAL_GPIO_WritePIN (GPIO_TypeDef* GPIOx ,uint16_t PIN_Number,uint8_t value );
-void MCAL_GPIO_WritePort (GPIO_TypeDef* GPIOx, uint8_t value );
+void MCAL_GPIO_WritePort (GPIO_TypeDef* GPIOx, uint16_t value );
 
 void  MCAL_GPIO_TogglePIN (GPIO_TypeDef* GPIOx ,uint16_t PIN_Number);
